This allows for the experience where the webappdown agent finds slow or completely down apps, and mitigates it, using the CPUAnalysis and AppCodeAnalysis agents

This pull request introduces new agent experiences for WebAppDown, CPUAnalysis, and AppCodeAnalysis.

src/Agent/FirstPartyAgent.FunctionApp/ApiController.cs: Added multiple new functions for handling configurations, alert details, and message processing.
src/Agent/FirstPartyAgent.Core/Services/ChatService/ChatProcessingService.cs: Enhanced chat processing service to support new agent modes and improved session handling.

----
#### AI description  (iteration 1)
#### PR Classification
Enhancement of agent functionalities and API improvements.

#### PR Summary
This pull request enhances the agent experience by refining the API controller and chat processing service functionalities. It includes improvements to alert handling and chat message processing.
- `ApiController.cs`: Improved alert configuration management and streamlined alert details retrieval and storage processes.
- `ChatProcessingService.cs`: Enhanced chat processing logic with better session management and agent loop handling, including improved message processing and integration with Teams.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

+++ src/Agent/Agent.Plugins/Models/MetricsModels.cs
@@ -22,7 +22,7 @@ namespace Agent.Plugins
 
     public sealed record ThreadTimeSeriesData(
         DateTime TimeStamp,
-        int ThreadCount);
+        double ThreadCount);
 
     public sealed record LatencySeriesData(
         DateTime TimeStamp,
